Fix project template test file
* Rename the test file so that it matches the unittest modules' default discovery pattern.
* Refactor the test_plugin test case so that it uses the implementation_resolver to test plugin registration
* Remove test_all_entities which doesn't work without an actual implementation
* Add necessary imports
